# Optimize nested writers during marshalling.