--- a/core/src/main/java/org/infinispan/protostream/annotations/impl/GeneratedMarshallerBase.java
+++ b/core/src/main/java/org/infinispan/protostream/annotations/impl/GeneratedMarshallerBase.java
@@ -3,11 +3,9 @@ package org.infinispan.protostream.annotations.impl;
 import java.io.IOException;
 
 import org.infinispan.protostream.ProtobufTagMarshaller;
-import org.infinispan.protostream.RandomAccessOutputStream;
+import org.infinispan.protostream.TagWriter;
 import org.infinispan.protostream.impl.BaseMarshallerDelegate;
 import org.infinispan.protostream.impl.Log;
-import org.infinispan.protostream.impl.RandomAccessOutputStreamImpl;
-import org.infinispan.protostream.impl.TagWriterImpl;
 
 /**
  * Base class for generated message marshallers. Provides some handy helper methods.
@@ -46,30 +44,8 @@ public class GeneratedMarshallerBase {
       if (ctx.depth() >= maxNestedMessageDepth) {
          throw log.maxNestedMessageDepth(maxNestedMessageDepth, message.getClass());
       }
-      try (NestedWriter nested = new NestedWriter(ctx, fieldNumber)) {
-         writeMessage(marshallerDelegate, nested.writer, message);
-      }
-   }
-
-   public static class NestedWriter implements AutoCloseable {
-      final RandomAccessOutputStream baos;
-      final TagWriterImpl writer;
-      private final ProtobufTagMarshaller.WriteContext ctx;
-      private final int fieldNumber;
-
-      public NestedWriter(ProtobufTagMarshaller.WriteContext ctx, int fieldNumber) {
-         this.ctx = ctx;
-         this.fieldNumber = fieldNumber;
-         this.baos = new RandomAccessOutputStreamImpl();
-         this.writer = TagWriterImpl.newNestedInstance(ctx, baos);
-      }
-
-      public TagWriterImpl getWriter() {
-         return writer;
-      }
-
-      public void close() throws IOException {
-         ctx.getWriter().writeBytes(fieldNumber, baos.getByteBuffer());
+      try (TagWriter nestedWriter = ctx.getWriter().subWriter(fieldNumber, true)) {
+         marshallerDelegate.marshall((ProtobufTagMarshaller.WriteContext) nestedWriter, null, message);
       }
    }
 }