--- a/core/src/main/java/org/infinispan/protostream/descriptors/WireType.java
+++ b/core/src/main/java/org/infinispan/protostream/descriptors/WireType.java
@@ -30,6 +30,8 @@ public enum WireType {
    public static final int FIXED_32_SIZE = 4;
    public static final int FIXED_64_SIZE = 8;
    public static final int MAX_VARINT_SIZE = 10;
+   // Similar to MAX_VARINT_SIZE except to be used for fields that are only ever used as Java primitive int values
+   public static final int MAX_INT_VARINT_SIZE = 5;
 
    /**
     * The lower 3 bits of the 32 bit tag are used for encoding the wire type.