--- a/core/src/main/java/org/infinispan/protostream/impl/ProtoStreamWriterImpl.java
+++ b/core/src/main/java/org/infinispan/protostream/impl/ProtoStreamWriterImpl.java
@@ -12,7 +12,6 @@ import java.util.Map;
 import org.infinispan.protostream.ImmutableSerializationContext;
 import org.infinispan.protostream.MessageContext;
 import org.infinispan.protostream.MessageMarshaller;
-import org.infinispan.protostream.RandomAccessOutputStream;
 import org.infinispan.protostream.TagWriter;
 import org.infinispan.protostream.descriptors.Descriptor;
 import org.infinispan.protostream.descriptors.FieldDescriptor;
@@ -507,11 +506,9 @@ final class ProtoStreamWriterImpl implements MessageMarshaller.ProtoStreamWriter
 
    private void writeMessage(FieldDescriptor fd, Object value, Class<?> clazz) throws IOException {
       BaseMarshallerDelegate marshallerDelegate = serCtx.getMarshallerDelegate(clazz);
-      RandomAccessOutputStream nestedBaos = new RandomAccessOutputStreamImpl();
-      TagWriterImpl nestedOut = TagWriterImpl.newNestedInstance(messageContext.out, nestedBaos);
-      marshallerDelegate.marshall(nestedOut, fd, value);
-      nestedOut.flush();
-      messageContext.out.writeBytes(fd.getNumber(), nestedBaos.getByteBuffer());
+      try (TagWriter nestedWriter = messageContext.out.subWriter(fd.getNumber(), true)) {
+         marshallerDelegate.marshall((TagWriterImpl) nestedWriter, fd, value);
+      }
    }
 
    private void writeGroup(FieldDescriptor fd, Object value, Class<?> clazz) throws IOException {
@@ -818,79 +815,75 @@ final class ProtoStreamWriterImpl implements MessageMarshaller.ProtoStreamWriter
       for (Map.Entry<? super K, ? super V> entry : map.entrySet()) {
          Object key = entry.getKey();
          validateElement(key, keyClass); // Protobuf 3 does not allow null keys
-         RandomAccessOutputStream nestedBaos = new RandomAccessOutputStreamImpl();
-         TagWriterImpl out = TagWriterImpl.newNestedInstance(messageContext.out, nestedBaos);
-         // Write the key as field 1
-         switch (md.getKeyType()) {
-            case BOOL -> out.writeBool(1, (Boolean) key);
-            case INT32 -> out.writeInt32(1, (Integer) key);
-            case INT64 -> out.writeInt64(1, (Long) key);
-            case FIXED32 -> out.writeFixed32(1, (Integer) key);
-            case FIXED64 -> out.writeFixed64(1, (Long) key);
-            case SINT32 -> out.writeSInt32(1, (Integer) key);
-            case SINT64 -> out.writeSInt64(1, (Long) key);
-            case SFIXED32 -> out.writeSFixed32(1, (Integer) key);
-            case SFIXED64 -> out.writeSFixed64(1, (Long) key);
-            case UINT32 -> out.writeUInt32(1, (Integer) key);
-            case UINT64 -> out.writeUInt64(1, (Long) key);
-            case STRING -> out.writeString(1, (String) key);
-            default ->
-                  throw new IllegalArgumentException("The Protobuf declared field type is not compatible with the written type : " + md.getFullName());
-         }
-         Object value = entry.getValue();
-         // Write the value as field 2
-         if (value == null) {
-            switch (md.getType()) {
-               case BOOL -> out.writeBool(2, false);
-               case INT32 -> out.writeInt32(2, 0);
-               case INT64 -> out.writeInt64(2, 0);
-               case FIXED32 -> out.writeFixed32(2, 0);
-               case FIXED64 -> out.writeFixed64(2, 0);
-               case SINT32 -> out.writeSInt32(2, 0);
-               case SINT64 -> out.writeSInt64(2, 0);
-               case SFIXED32 -> out.writeSFixed32(2, 0);
-               case SFIXED64 -> out.writeSFixed64(2, 0);
-               case UINT32 -> out.writeUInt32(2, 0);
-               case UINT64 -> out.writeUInt64(2, 0);
-               case DOUBLE -> out.writeDouble(2, 0);
-               case FLOAT -> out.writeFloat(2, 0);
-               case STRING -> out.writeString(2, "");
-               case BYTES -> out.writeBytes(2, new byte[0]);
-               case ENUM -> out.writeEnum(2, 0);
+         try (TagWriter out = messageContext.out.subWriter(fieldNumber, true)) {
+            // Write the key as field 1
+            switch (md.getKeyType()) {
+               case BOOL -> out.writeBool(1, (Boolean) key);
+               case INT32 -> out.writeInt32(1, (Integer) key);
+               case INT64 -> out.writeInt64(1, (Long) key);
+               case FIXED32 -> out.writeFixed32(1, (Integer) key);
+               case FIXED64 -> out.writeFixed64(1, (Long) key);
+               case SINT32 -> out.writeSInt32(1, (Integer) key);
+               case SINT64 -> out.writeSInt64(1, (Long) key);
+               case SFIXED32 -> out.writeSFixed32(1, (Integer) key);
+               case SFIXED64 -> out.writeSFixed64(1, (Long) key);
+               case UINT32 -> out.writeUInt32(1, (Integer) key);
+               case UINT64 -> out.writeUInt64(1, (Long) key);
+               case STRING -> out.writeString(1, (String) key);
                default ->
                      throw new IllegalArgumentException("The Protobuf declared field type is not compatible with the written type : " + md.getFullName());
             }
-         } else {
-            switch (md.getType()) {
-               case BOOL -> out.writeBool(2, (Boolean) value);
-               case INT32 -> out.writeInt32(2, (Integer) value);
-               case INT64 -> out.writeInt64(2, (Long) value);
-               case FIXED32 -> out.writeFixed32(2, (Integer) value);
-               case FIXED64 -> out.writeFixed64(2, (Long) value);
-               case SINT32 -> out.writeSInt32(2, (Integer) value);
-               case SINT64 -> out.writeSInt64(2, (Long) value);
-               case SFIXED32 -> out.writeSFixed32(2, (Integer) value);
-               case SFIXED64 -> out.writeSFixed64(2, (Long) value);
-               case UINT32 -> out.writeUInt32(2, (Integer) value);
-               case UINT64 -> out.writeUInt64(2, (Long) value);
-               case STRING -> out.writeString(2, (String) value);
-               case BYTES -> out.writeBytes(2, (byte[]) value);
-               case ENUM -> out.writeEnum(2, ((Enum<?>) value).ordinal());
-               case MESSAGE -> {
-                  // FIXME: there is too much nesting here. We can definitely improve things
-                  BaseMarshallerDelegate<V> marshallerDelegate = serCtx.getMarshallerDelegate(valueClass);
-                  RandomAccessOutputStream mapValueBaos = new RandomAccessOutputStreamImpl();
-                  TagWriterImpl mapValueOut = TagWriterImpl.newNestedInstance(out, mapValueBaos);
-                  marshallerDelegate.marshall(mapValueOut, md, (V) value);
-                  mapValueOut.flush();
-                  out.writeBytes(2, mapValueBaos.getByteBuffer());
+            Object value = entry.getValue();
+            // Write the value as field 2
+            if (value == null) {
+               switch (md.getType()) {
+                  case BOOL -> out.writeBool(2, false);
+                  case INT32 -> out.writeInt32(2, 0);
+                  case INT64 -> out.writeInt64(2, 0);
+                  case FIXED32 -> out.writeFixed32(2, 0);
+                  case FIXED64 -> out.writeFixed64(2, 0);
+                  case SINT32 -> out.writeSInt32(2, 0);
+                  case SINT64 -> out.writeSInt64(2, 0);
+                  case SFIXED32 -> out.writeSFixed32(2, 0);
+                  case SFIXED64 -> out.writeSFixed64(2, 0);
+                  case UINT32 -> out.writeUInt32(2, 0);
+                  case UINT64 -> out.writeUInt64(2, 0);
+                  case DOUBLE -> out.writeDouble(2, 0);
+                  case FLOAT -> out.writeFloat(2, 0);
+                  case STRING -> out.writeString(2, "");
+                  case BYTES -> out.writeBytes(2, new byte[0]);
+                  case ENUM -> out.writeEnum(2, 0);
+                  default ->
+                        throw new IllegalArgumentException("The Protobuf declared field type is not compatible with the written type : " + md.getFullName());
                }
-               default ->
-                     throw new IllegalArgumentException("The Protobuf declared field type is not compatible with the written type : " + md.getFullName());
+            } else {
+               switch (md.getType()) {
+                  case BOOL -> out.writeBool(2, (Boolean) value);
+                  case INT32 -> out.writeInt32(2, (Integer) value);
+                  case INT64 -> out.writeInt64(2, (Long) value);
+                  case FIXED32 -> out.writeFixed32(2, (Integer) value);
+                  case FIXED64 -> out.writeFixed64(2, (Long) value);
+                  case SINT32 -> out.writeSInt32(2, (Integer) value);
+                  case SINT64 -> out.writeSInt64(2, (Long) value);
+                  case SFIXED32 -> out.writeSFixed32(2, (Integer) value);
+                  case SFIXED64 -> out.writeSFixed64(2, (Long) value);
+                  case UINT32 -> out.writeUInt32(2, (Integer) value);
+                  case UINT64 -> out.writeUInt64(2, (Long) value);
+                  case STRING -> out.writeString(2, (String) value);
+                  case BYTES -> out.writeBytes(2, (byte[]) value);
+                  case ENUM -> out.writeEnum(2, ((Enum<?>) value).ordinal());
+                  case MESSAGE -> {
+                     // FIXME: there is too much nesting here. We can definitely improve things
+                     BaseMarshallerDelegate<V> marshallerDelegate = serCtx.getMarshallerDelegate(valueClass);
+                     try (TagWriter nestedWriter = out.subWriter(2, true)) {
+                        marshallerDelegate.marshall((TagWriterImpl) nestedWriter, md, (V) value);
+                     }
+                  }
+                  default ->
+                        throw new IllegalArgumentException("The Protobuf declared field type is not compatible with the written type : " + md.getFullName());
+               }
             }
          }
-         out.flush();
-         messageContext.out.writeBytes(fieldNumber, nestedBaos.getByteBuffer());
       }
    }