… core/src/test/java/org/infinispan/protostream/ProtobufUtilTest.java