Fixes #427

This PR introduces a new method to TagWriter and Encoder to allow those classes control the TagWriter and Encoder that are created whenever a sub message is found. It removes all manual creations of TagWriters in those places and replaces with calls to the new method.

The first commit does that refactoring and by default has all Writer/Encoders use OutputStreamRandomAccessEncoder as was done prior. The big difference is the new encoders will cache previously used sub encoders so they will not have to recreate them for multiple nested messages at the same depth.

The second commit takes this one step further by introducing a move method to the RandomAccessOutputStream by allowing it to move bytes within the same stream. This allows us to not have to create any additional byte[] and instead it reuses the same one from the parent.

The third commit is just one to remove some unneeded computations for var ints where it was finding the exact size to ensure the byte[] was large enough. In that case there is no real reason to incur the true cost and just assume 5 for the worst case. In many cases this will not cause a resize and the cases when it does it most likely would have had to resize for the next value anyways.

The fourth commit takes the second commit a step further by actually reusing the same Encoder instance for all nested messages. Instead the previous sizes are stored in an int[] and when close is invoked it takes the last size value to write to it, further reducing allocation costs.