@wburns 2025-02-26
Running the most up to date benchmark at https://github.com/infinispan/infinispan-benchmarks/tree/main/protostream I got the following numbers compared with main and the new PR. In this case it is ns/op so a lower value is better. In this case the new code is anywhere from 16% to nearly 3x faster.

MAIN

Benchmark | (byteArrayOrStream) | (userByteArraySize) | Mode | Cnt | Score |   | Error | Units
-- | -- | -- | -- | -- | -- | -- | -- | --
ProtostreamBenchmark.testMarshallAddress | TRUE | N/A | avgt | 10 | 82.104 | ± | 0.346 | ns/op
ProtostreamBenchmark.testMarshallAddress | FALSE | N/A | avgt | 10 | 66.607 | ± | 0.423 | ns/op
ProtostreamBenchmark.testMarshallIracMetadata | TRUE | N/A | avgt | 10 | 214.432 | ± | 1.035 | ns/op
ProtostreamBenchmark.testMarshallIracMetadata | FALSE | N/A | avgt | 10 | 192.916 | ± | 0.73 | ns/op
ProtostreamBenchmark.testMarshallUser | TRUE | 10 | avgt | 10 | 479.489 | ± | 1.669 | ns/op
ProtostreamBenchmark.testMarshallUser | TRUE | 8096 | avgt | 10 | 1881.338 | ± | 7.82 | ns/op
ProtostreamBenchmark.testMarshallUser | FALSE | 10 | avgt | 10 | 455.762 | ± | 2.609 | ns/op
ProtostreamBenchmark.testMarshallUser | FALSE | 8096 | avgt | 10 | 1028.067 | ± | 1.778 | ns/op

NEW

Benchmark | (byteArrayOrStream) | (userByteArraySize) | Mode | Cnt | Score |   | Error | Units
-- | -- | -- | -- | -- | -- | -- | -- | --
ProtostreamBenchmark.testMarshallAddress | TRUE | N/A | avgt | 10 | 69.325 | ± | 0.177 | ns/op
ProtostreamBenchmark.testMarshallAddress | FALSE | N/A | avgt | 10 | 54.123 | ± | 0.156 | ns/op
ProtostreamBenchmark.testMarshallIracMetadata | TRUE | N/A | avgt | 10 | 99.783 | ± | 0.318 | ns/op
ProtostreamBenchmark.testMarshallIracMetadata | FALSE | N/A | avgt | 10 | 97.352 | ± | 1.037 | ns/op
ProtostreamBenchmark.testMarshallUser | TRUE | 10 | avgt | 10 | 276.391 | ± | 0.816 | ns/op
ProtostreamBenchmark.testMarshallUser | TRUE | 8096 | avgt | 10 | 1473.529 | ± | 2.012 | ns/op
ProtostreamBenchmark.testMarshallUser | FALSE | 10 | avgt | 10 | 242.012 | ± | 0.663 | ns/op
ProtostreamBenchmark.testMarshallUser | FALSE | 8096 | avgt | 10 | 352.997 | ± | 1.907 | ns/op

The tests done stress specific things

Address - most basic (has no nested messages other than the wrapped message around it)
IracMetadata - has many nested and sub nested messages
User - has some nested messages and a very large byte[] in one case

Also to note the byteArrayOrStream method is if the final destination is to write the message to a byte[] and then copy it into another RandomByteArrayOutpuStream or if the message is directly copied into a RandomByteArrayOutputStream. This explains why the User test with a large array is so much faster with the stream variant.