@wburns 2025-02-27
> > Updated, I will squash the commits later today if no more comments.
> 
> Looks good in general, but there are a lot of test failures according to the GH check.

Should be fixed now, when I added the maxDepth the arguments got switched in one case 🤦